Add Android Beam for Issues
The Beam API calls are wrapped in the 'Beam' class, so that the
classloaders of earlier devices don't throw exceptions.

To get Beam to work, you need to specify in your activity code that
you're going to broadcast a Beam message, and in your manifest that
your activity is interested in receiving particular Beam messages. We
filter here by specifying the "application/vnd.github.repo.issue+json"
mime type for messages.

The integration test exposed an interesting bug in robotium - if you
decide to skip the body of a test (eg because beam isn't available for
the device under test), the teardown method will hang on the robotium
method finishOpenedActivities() - it has a lethal little loop that
doesn't die:

https://github.com/jayway/robotium/blob/robotium-3.1/robotium-solo/src/main/java/com/jayway/android/robotium/solo/ActivityUtils.java#L120..L122


NFC & Beam documentation:
http://developer.android.com/guide/topics/nfc/nfc.html