Fix 'refresh' button on action bar so that it forces a fresh load
Previously cached data would be used, which made the refresh very quick
but wouldn't provide an update unless the sync adapter had run.

The reload button now causes a 'force-reload' parameter to be passed to
the loader - if the loader checks this using the 'isForcedReload()'
method it can force a reload accordingly. Only done for the RepoList
fragment at the moment.