Check for repository completeness before configuring pager
Previously a repository opened from the news feed with a
valid owner would be missing an issues page even if issues
were enabled.  This was because the avatar URL was present
but the repository details were not checked for completeness.

Add a simple checking mechanism to know whether we are
dealing with a repository model class returned from an
API call or the database cache.