Don't refresh list when item is clicked
This was originally done to mark the repository
as recently viewed in the list but can cause
confusion since the item appears to disappear from
the current position in the list that is visible
since it moves to the top area which is most likely
not visible.

The repository will still appear in the recent list
the next time the activity is created.

Closes issue #71