Translate "news" as "actualizaciones"
The literal translation for "news" would be "noticias" but it sound a
bit awkward in this case, since in GitHub's case these are more like
updates than proper news.
An anternative way to translate it would be "eventos" (meaning events),
but I feel "actualizaciones" (meaning updates) is a better term.

Signed-off-by: Jose M. Chumo Mata <jmchuma@gmail.com>