Remove use of nested adapters in HomeDropdownListAdapter
Previously two adapters were used, one for regular view
and one for the dropdown view.  This consolidates them
into a single adapter that contains items for all the
orgs and the actions that display after the orgs.

This allows views to be reused and remove the ugliness
of deferring from one parent adapter to two different
child adapters depending on the position