Load PDFs in WebView through Google Docs
URLs for the raw PDF are loaded in WebView using Google Docs
embedded viewer.
There is a button on the page to download the file and view
in an external editor.

The alternative to this is to immediately download the file to
the SD card then launch an intent to view the downloaded PDF or
add a button to the action bar to download the file.
This would require additional permissions to write to external storage.