Translation a word "repository"
In the community of Russian programmers decided to translate "repository" as a "репозиторий", and "хранилище" as a storage.  	
You can see the context of articles on the results of search queries popular in Russia IT .

http://habrahabr.ru/search/?q=репозиторий
http://habrahabr.ru/search/?q=хранилище

Google Translate will help to understand at least a rough gist of the use of these words.