Replace existing icons with new material variants where appropriate
Was able to remove some that were duplicates or already existed in appcompat.

This doesn't touch the icons in nodpi

Also removed one instance (delete history) where it was used in a menu but never actually shown. This allows us to use a regular grey one for the repo dialogs and now have to worry about adding a second one that's white to match the actionbar.