Move comment edit/delete controls to a popup menu with overflow
This eliminates the need to have a second set of delete/edit icons for placing in the comment, and also makes it a little less redundant by moving them into a popup menu. When you have menu items of sorts in list items, it's more conventional to have them hidden unless the user needs them to avoid cluttering the view.