Add delete repository feature. Fix #376
This makes it possible to delete a repo just like in the web.
This can also be used as a baseline to migrate existing API consumption
to Retrofit.

Add new authorization request for delete token

If token with "delete_repo" doesn't exist, then
it will create a new one.

Refresh list after deletion